backend/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
faster-whisper
torch
sentencepiece
argostranslate
pydantic
python-dotenv
